+++ CHANGELOG.md
@@ -5,6 +5,12 @@
 - feat: support `--exclude` and `--exclude-dependents` flags
   ([#1363](https://github.com/pulumi/actions/pull/1363))
 
+- feat: verbosity and debug inputs to enable debug options
+  ([#1046](https://github.com/pulumi/actions/pull/1046))
+
+- fix: update "user agent" string to pulumi/actions@v6
+  ([#1357](https://github.com/pulumi/actions/pull/1367))
+
 ---
 
 ## 6.2.0 (2025-03-05)